% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vakie3.R
\docType{data}
\name{data_vakie3}
\alias{data_vakie3}
\title{Population data by municipality}
\format{
A data frame with the following columns:
\describe{
\item{tunnus}{Municipality code}
\item{nimi}{Municipality name}
\item{Maakunta}{Province}
\item{value}{Population count}
}
}
\source{
Statistics Finland
}
\usage{
data_vakie3
}
\description{
Population development and projections for Finnish municipalities,
covering observed values (2000–2019) and projections (2020–2040).
}
\keyword{datasets}
