% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_category.R
\docType{data}
\name{pop_category}
\alias{pop_category}
\title{Population development by municipality category}
\format{
A data frame in long format with 360 rows and 4 variables:
\describe{
\item{year}{Year}
\item{category}{Urban–rural municipality category}
\item{value}{Population change in persons}
\item{variable}{Demographic component of population change}
}
}
\source{
Statistics Finland.
\emph{Migration and population change database}.
\url{https://stat.fi/fi/palvelut/tilastodatapalvelut/tilastotietokannat/vaestotilastopalvelu/muuttoliike-ja-vaestonmuutokset}
}
\usage{
pop_category
}
\description{
Components of population change by urban-rural municipality category.
}
\keyword{datasets}
